Implement long for 16 digit card number

+++ app/src/main/java/VendingMachine/PayCard.java
@@ -111,11 +111,11 @@ public class PayCard extends Page {
      */
     public static boolean checkCardNumber(String cardNumber) {
 
-        int number;
+        long number;
 
         // Checks if integer
         try {
-            number = Integer.parseInt(cardNumber);
+            number = Long.parseLong(cardNumber);
         } catch (Exception e) {
             return false;
         }
